requirements-dev.txt
# Development dependencies
-r requirements.txt

# Testing
pytest==7.4.3
pytest-cov==4.1.0
pytest-flask==1.3.0

# Code formatting and linting
black==23.9.1
flake8==6.1.0
isort==5.12.0

# Development tools
python-dotenv==1.0.0
